Automated g4 rollback of changelist 129603738.
*** Reason for rollback ***

Breaks some tests as this CL introduces a dependency for goog.dom.animationFrame which assumes there's always a window.

*** Original change description ***

Listens to the ORIENTATIONCHANGE event in modalpopup.js.

Adds a hint in animationframe.js to tell the compiler the type of this so that it can be used within closure.

RELNOTES: n/a

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129624202