Fail goog.getCssName when className starts with a dot.
This was suggested in CL 125557257.

When passing a css selector to goog.getCssName like so
.findFirst(goog.getCssName('.quantumWizButtonPaperbuttonLabel'))
this works in test mode.

However, with a mapping enabled in production, it starts failing, the right usage being
.findFirst('.' + goog.getCssName('quantumWizButtonPaperbuttonLabel'))

RELNOTES: Failing on goog.getCssName if a dot is passed as className.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=129987935