RELNOTES[INC]: sanitize unknown tags as <span> instead of <div>. For consistency, sanitizeToDomNode now wraps the output in a <span> tag instead of a <div>.
The current implementation replaces unknown tags (i.e tags that are neither in the whitelist nor in the blacklist) with <div> tags, but that can lead to malformed HTML when a <div> tag appears inside an inline element. Browsers will attempt to fix this by closing and opening new tags, resulting in spurious elements. This CL switches to <span> tags, which are inline and are not corrected by the browser.

Note that the result could still be invalid HTML! The actual spec is way more complex than this. We can only hope that browsers do not attempt to fix the output when out-of-place <span> tags are involved. Capturing the full HTML5 spec is out of scope.

The CL also updates the documentation for sanitizeToDomNode (it claimed to return a NodeList) and wraps its output in a <span> instead of a <div> for consistency.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130340276